Rename presenter
Getting ahead of ourselves!

We intend to rename all our models and services to reflect how they are referred to in the UI, or within the team. At the moment though we are using the database as our guide but we know this is the most 'wrong' of how things are now referenced in the service.