Refactor CreateEventService for bill runs (#73)
https://eaflood.atlassian.net/browse/WATER-3831

We've created the model and a service wrapper for creating events. But at this time we are only interested in creating an `event` when a new bill run is initialised. That means some fields will always be set with a certain value.

Also, based on reviewing existing events created by the legacy code the `metadata` holds a formatted version of a bill run. So, we need to handle that.

All this means we'd be better placed to refactor the existing service to focus on new bill-run events.

** Notes

- Add migrations for missing fields
  There are columns on the `regions` and `billing_batches` tables that the existing logic extracts data from that isn't present in our test db. So, we add alter migrations for the missing fields and update our helpers accordingly.

- Add presenter needed to format metadata content
  The metadata field in an event record doesn't just contain a dump of the BillingBatch as JSON. Instead, it is a specifically formatted object. So, we need something that will convert the `BillingBatchModel` instance into the content we need.

- Add source as a default in BillingBatchHelper
  Source is populated when a billing batch event is created for real. So, we ensure it's populated in our helper to reflect that.