Refactor CreateBillingBatchService
https://eaflood.atlassian.net/browse/WATER-3854

To support our work on adding an endpoint that will allow us to create a new SROC bill run we need to refactor the existing `CreateBillingBatchService` a little bit!

Some of the stuff it's currently defaulting we're now going to pass in. We also need it to bring back the instance with `.region` populated. This is to support creating the 'new' billing batch event record.