Add user to CreateBillRunValidator
https://eaflood.atlassian.net/browse/WATER-3854

As part of initialising a new bill run, we create an 'event' record. It requires the `issuer` field to be populated with the email address of the user requesting the new bill run.

So, we'll need to include the current user's email as part of our request. This change updates the validator to include it.