Add new initialise Billing Batch service
At the start of a new bill run we need to

- create the initial billing batch record
- create the associated event record
- send a request to the SROC Charging Module API and link it to our billing batch

With those 3 things done we can return control back to the requester (which will be the UI) whilst we get on with the rest of the bill run process.

This adds the service that will handle those 3 steps. We're waiting on additional functionality before we can do the last one. But for now steps 1 & 2 are done and we'll have a home for step 3 when it's ready!