Update "create bill run" endpoint to create a bill run (#56)
* Update "create bill run" endpoint to create a bill run

https://eaflood.atlassian.net/browse/WATER-3854

When we set up our "create bill run" endpoint, we had it simply respond to requests with a dummy response. We now update it to actually create a bill run and an event, and return the required details in the same format as the dummy response.

* Add `idColumn` static method to `BillingBatchModel`

As we start to look at creating entries in the db, we run into an issue when Objection errors because it can't find an `id` column. On investigation it turns out that if the primary id column isn't called `id`, it needs to be specified in the model. We therefore add an `idColumn` static method to our billing batch model

* Initial pass at inserting bill run in db

We update our `createBillRun` controller to create a record for the bill run in the db. We initially hardcode the financial year fields just to get it working, with the intention of pulling these through properly from the request in a later commit

* Move bill runs controller into folder

We try and follow that the path you see in the url matches where the controller sits in the repo.

* Add presenter for formatting response

We need something that handles formatting the response for the POST `/bill-runs` endpoint. We use what we did in https://github.com/DEFRA/water-abstraction-ui/pull/2247 as a spec.

* Add new initialise Billing Batch service

At the start of a new bill run we need to

- create the initial billing batch record
- create the associated event record
- send a request to the SROC Charging Module API and link it to our billing batch

With those 3 things done we can return control back to the requester (which will be the UI) whilst we get on with the rest of the bill run process.

This adds the service that will handle those 3 steps. We're waiting on additional functionality before we can do the last one. But for now steps 1 & 2 are done and we'll have a home for step 3 when it's ready!

* Refactor bill-runs controller to use new service

Now we've created the InitiateBillingBatchService we refactor the bill runs controller to use it.

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>