Restructure models by type and scheme
https://github.com/DEFRA/water-abstraction-team/issues/69

Our next step in refactoring the models to deal with legacy issues is to identify them by type and scheme. Admittedly, we know we are just dealing with the one type (legacy) and scheme (water). But more will be coming and this helps set the groundwork for implementing base classes that allow us to start hiding some of the issues with the existing DB.

So, we move the existing models into a new `models/water` folder and we add a new `LegacyModel` which all our existing classes will inherit from.