Restructure the base models
This means adding 2 new base models, one extending from the other with `BaseModel` as the root parent.

This also means we can move out of BaseModel functionality we only added to support working with the legacy service and it's various schemas. When we get to add our own models, they won't need to inherit this cruft.

Whilst doing this we were able to make some changes to the whole declaring the schema name solution. The original was made in 2016, so for example `RelatedQueryBuilder` is defunct (should have spotted that before 🤦). And as we only need this in the `LegacyBaseModel` we no longer assign our custom QueryBuilder to Objection's root Model `QueryBuilder` property.

Finally, we added some documentation around all this so if we ever have to look at it again we'll have a clue what what it's all about.