Replace `standard` with `standardx`
StandardJS covers 95% of our linting needs in one simple package. However there are some things it has no opinion on which we feel strongly about -- for example, ensuring parameters are always enclosed in brackets (even in cases where a single parameter's brackets are optional).

We'd like to be able to flag these things up during linting and `standardx` sems like the perfect package -- a drop-in replacement for `standard` which allows extra rules to be added. This means we can extend our linting without the pain of installing full-fat ESLint.