Get `standardx` actually working
Dropping in `standardx` in place of `standard` didn't immediately work!

The first problem was an error `Failed to load config "standard" to extend from`. There was no consensus online how to fix this, some people suggested installing a bunch of ESLint modules, however we found that simply reinstalling `standard` did the trick.

The second problem was numerous `Parsing error: Invalid ecmaVersion` errors. This seems to be an incompatibility between `standardx` and `eslint`, where `eslint` defaults to ecmaVersion 13 but `standardx` isn't compatible with this. Enforcing ecmaVersion 12 in our `package.json` resolved this.