Determine if charge version is 'billed'
https://eaflood.atlassian.net/browse/WATER-3868

We also need to know whether a charge version has already been ‘billed’ in the current period. If so we’ll have to ensure it’s credited as part of the bill run before we recalculate the debit.

For each charge version that is a candidate for billing, we need to determine if it was included in a `billing_batch` for the current period.

For now, the result of this check should be included in the `/check/supplementary` endpoint results, against each charge version. That way we can verify if it is working correctly.