Request new bill run in Charging Module API
https://eaflood.atlassian.net/browse/WATER-3833

As part of the bill run process, we need to send a request to the Charging Module API to create a new bill run.

To do this we'll need to handle authenticating with the Charging Module's AWS Cognito service.

The ID it returns needs to be stored against our billing_batch record.

We also need to be able to handle 4xx and 5xx responses.