Make timestamps consistent at model layer
https://github.com/DEFRA/water-abstraction-team/issues/69

We identified that there are inconsistencies in the legacy database, timestamps in the tables being one of them. We don't want these issues making a mess of our code. So, until we're ready for some full scale data migration we're going to deal with them in our model layer.

Timestamps being different across tables is the main inconsistency we noted and this change deals with it.