Add authentication and fix request body
We add the simplest method of authentication which is to call our `ChargeModuleTokenService` every time a request is made then add the returned token to our request headers.

While doing this we realised that we should have been passing the region and rulest in `options.body` and not simply as part of `options`.