Add test for translations in the LegacyBaseModel
We didn't think it necessary to worry about how the data actually looks in the DB. So, we call the overridden methods directly that Objection.js will use.

They'll be picked up just by the act of using the helpers in our other unit tests that do write to the DB. Plus going lower felt like we're starting to test Objection.