Update rest of models with their translations
Looking at this change there is an argument to move the translations into `WaterBaseModel` and then just override the exceptions like `EventModel`. But we have an eye to also using this logic to make the ID property for each model _actually_ `.id`. And that will be different in each which means every model is going to need to set `translations ()`. So, it'd better to keep the translations together in one place for each model, rather than splitting them.