Fix issue with implementation
A number of tests started failing, all with the issue that `date_created` could not be null. The source was the helpers but we're generally not setting those fields in the helpers. We expect the DB to default them??

Obviously it was our new code, and we tracked it down to the fact it was _always_ applying the translation to the model JSON, even if the property didn't exist.

For example, `RegionHelper.add()` adds a new region but doesn't set `date_created` as a property on the model. The DB should be default it on insert. But we have a translation for it. The `$formatDatabaseJson()` was seeing the translation and calling

```javascript
json[translation.database] = json[translation.model]
```

So, `dateCreated` doesn't exist in the JSON, but we've added the property. Then we set it to `undefined`, because we're again referring to something that wasn't added to the model instance. Result, Objection.js tries to INSERT `date_created` as `null`.

We've updated the logic in $formatDatabaseJson() and $parseDatabaseJson(), just to be on the safe side, to only consider translations that exist on the json object being modified. The javascript _in_ operator returns true or false whether the named property exists in the object. An undefined or null property still exists so we get `true` back in those cases, which is what we want as there may be times we are intending to set a field to null.