Add a test to demonstrate queries don't work
In the course of implementing this change we've realised that we cannot use standard property names in queries. We still have to use the legacy column names.

We've chased down every possible option we could find to support both working with instances _and_ queries but it's no use.

We'll fully document what we found and tried in the code when we start adding documentation.