Fix tests breaking due to `require` order
We previously found that our tests were breaking if `InitiateBillingBatchService` brought in `ChargingModuleCreateBillRunService` ahead of certain other services. After investigation we realised this was because the chain of requirements ended up bringing in one of our config files prior to `dotenv` loading in the environment variables. We've resolved to fix this by adding `dotenv` to all our config files, which we will do in a separate PR; in the meantime we can fix this simply by changing the order that `CreateBillRunService` brings in its requirements, moving the services to after the config (which we feel is preferable to having `InitiateBillingBatchService` require its services in a non-alphabetical order)