Make timestamps consistent at model layer (#85)
https://github.com/DEFRA/water-abstraction-team/issues/69

We identified that there are inconsistencies in the legacy database, timestamps in the tables being one of them. We don't want these issues to make a mess of our code. So, until we're ready for some full-scale data migration we're going to deal with them in our model layer.

Timestamps being different across tables is the main inconsistency we noted and this change deals with it.

We use [Objection.js](https://vincit.github.io/objection.js/api/model/instance-methods.html#parsedatabasejson) `$parseDatabaseJson()` and `$formatDatabaseJson()` to present a consistent API in our models whilst still working with the DB. This is built into the `LegacyBaseModel` with every child model expected to include a `translations()` getter that will be used by both. There is an argument to move `translations()` into `WaterBaseModel` and then just override it for exceptions like `EventModel`. But we have an eye to also use this logic to make the ID property for each model _actually_ `.id`. And that will be different in each which means every model is going to need to set `translations ()`. So, it'd be better to keep the translations together in one place for each model, rather than splitting them.

> WARNING! Queries cannot use the 'code' version of a column name. They must use the database version. We've included a unit test to demonstrate this and documentation to explain the reasons why.