Add new GeneralLib module
We need to get the current date as a ISOString to pass to PostgreSQL for reasons explained in the comments.

This doesn't feel like reason enough to build a standalone `TimeLib` or `PostgreSqlLib`. But we do need to keep an eye on it and ensure it doesn't become a 'kitchen-drawer' of methods.