Request new bill run in Charging Module API (#83)
https://eaflood.atlassian.net/browse/WATER-3833

As part of the bill run process, we need to send a request to the Charging Module API to create a new bill run. We therefore update `InitiateBillingBatchService` to call the recently-created `ChargingModuleCreateBillRunService` to do this, and store the returned id in the `billing_batch` record as `externalId`.

In terms of error handling we simply rely on the individual stages throwing an error, which we then catch in the controller. The one exception is the result of `ChargingModuleCreateBillRunService`, which always returns a `{ succeeded: ..., response: ... }` tuple. We therefore check for failure and throw an error (which as before is caught by the controller).

The controller takes the caught error and reports a `500` error back. Note that Boom does not expose the message of a `500` error in its response; if we want to see the message we would need to check the `system` logs.