Cache and reuse Charging Module JWT token
https://eaflood.atlassian.net/browse/WATER-3867

[We previously built our first link to the SROC Charging Module API](https://eaflood.atlassian.net/browse/WATER-3833). That included authenticating with the Module's AWS Cognito service to then get a JWT which needs to be sent with all subsequent requests.

These tokens expire after an hour by default, at which point another one needs to be requested. Because we'll need to send thousands of requests to the charging module during a bill run it's important we can reuse a token as much as possible and avoid the overhead of requesting new ones.

This means we need to add to`water-abstraction-system` the ability to cache tokens and reuse them as part of communicating with the Charging Module API.