Fix service following Knex update
A recent change and test appear to be failing the build after this update to Knex. Checking the Knex changelog we found this

- [Bug fix: PostgreSQL: Fix Malformed array literal 2.4.0 Regression #5439](https://github.com/knex/knex/blob/master/CHANGELOG.md#241---18-january-2022)

'Array' you say!? [The PR](https://github.com/knex/knex/pull/5439) links to various issues which all seem to have JSONB and arrays as primary concerns.

TBH, we don't have the time to properly dig into why it worked before but now doesn't. But as this was a fix to a regression the behaviour is unlikely to change in Knex so we need to code for it.

This change resolves the issue for us. Looking in the DB the column is back to holding an empty array (without the fix Knex's change made it an empty JSON object) which matches what the legacy system is doing.