Refactor cache plugin and call server method from `/get` route
After further reading we come to the conclusion that the simplest way to cache a token is to create a server method. By providing cache config options with we define this method, it will automatically cache the result when called and serve this cached result until the expiry time has been reached.

We therefore refactor our code so that our `CachePlugin` is now the `ChargingModuleTokenCachePlugin`, which registers `server.methods.getChargingModuleToken()`. Currently this simply returns the current time, cached for 10 seconds. But it's enough to be sure that this approach to caching works!