Override expiry time
Currently our requested tokens expire in an hour. We don't expect this to change, but we don't want to make the assumption just in case! We therefore change the cache expiry time to default to 1 day, but add an override in the main method to manually set it to 10 seconds