Create `ChargingModuleTokenCachePlugin` to cache Charging Module JWT token (#91)
https://eaflood.atlassian.net/browse/WATER-3867

[We previously built our first link to the SROC Charging Module API](https://eaflood.atlassian.net/browse/WATER-3833). That included authenticating with the Module's AWS Cognito service to get a JWT which needs to be sent with all subsequent requests.

These tokens expire after an hour by default, at which point another one needs to be requested. Because we'll need to send thousands of requests to the charging module during a bill run it's important we can reuse a token as much as possible and avoid the overhead of requesting new ones.

We do this by creating a new plugin `ChargingModuleTokenCachePlugin` which registers a server method to get the token. This takes advantage of the fact that hapi has built-in caching for these server methods.

Our new `server.methods.getChargingModuleToken()` will return the cached `token` object if one exists, or retrieve one if it doesn't (either because one hasn't been fetched yet, or because the cached token has expired).

Although we don't expect the expiry time of the tokens to change from the current value of an hour, we still use the returned `expiresIn` value to determine when the cache should expire. However we set it to be a minute less than the returned value, just to avoid the (admittedly unlikely!) case that a cached token is retrieved which then expires prior to its use.

Note that we haven't yet updated `CreateBillRunService` to use this new server method; this will be done in another PR.

We may also want to consider changing the cache options to use Redis instead of in-memory caching but again, this can be done in a future PR.