Add comments to AbstractionBillingPeriodService (#93)
https://eaflood.atlassian.net/browse/WATER-3834

The `AbstractionBillingPeriodService` is doing some hairy stuff with dates. With no context as to the reasons why it becomes at risk of becoming a confusing mess like the legacy code.

So, this attempts to add some context and examples to the service to help explain what it is doing, but more importantly why.