Support new queued bill run status
https://eaflood.atlassian.net/browse/WATER-3877

When a supplementary bill run is started, we'll be creating both a PRESROC and SROC one at the same time.

The PRESROC one will then begin `processing` (labelled as BUILDING in the UI) whilst the SROC one is `queued` waiting to be processed.

At the moment though, all we can do is initialise the SROC bill run with a status of `processing` as that is all that is available to us. In the UI this will give the impression both are building when they're not.

We've added a new `queued` status to [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service) and support for it in [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui). This change updates changes the default status we apply to new bill runs to be `queued`.