Fix handling of Charging Module errors
https://eaflood.atlassian.net/browse/WATER-3833

While diagnosing an issue with Charging Module authorisation, we realised that we weren't correctly passing back any error messges it returned. It turned out we hadn't correctly grasped the format that errors come back -- while the status code and status error are returned as part of the response, the actual error we want to display (eg. `Invalid JWT: Token format not valid`) is returned in the response body (along with the status code and error). We therefore update how we return errors.