Reformat Charging Module errors in `InitiateBillingBatchService`
We're now getting errors back from `CreateBillRunService` in the expected format, but just to cover the eventuality that an error doesn't return a message in the body, we reformat the thrown error to always return the status code and error (eg. `401` and `Unauthorized`), and optionally return the message if it exists (eg. `Invalid JWT: Token format not valid`)