Fix handling of Charging Module errors (#96)
https://eaflood.atlassian.net/browse/WATER-3833

While diagnosing an issue with Charging Module authorisation, we realised that we weren't correctly passing back any error messages it returned.

It turned out the format that errors come back in didn't quite match what we expected -- while the status code and status error (eg. `401` and `Unauthorized`) are returned as part of the response, the error message we want to display (eg. `Invalid JWT`) is returned in the response body in JSON format (which also contains the status code and error).

We therefore update how `CreateBillRunService` returns these errors by parsing the body, and in the event of an unsuccessful request we return it as-is in the response (whereas a successful request still extracts the `billRun` object from the body and returns that as the response). This gives us the status code and status error along with any returned error message.

We then amend how `InitiateBillingBatchService` formats these errors. We previously intended errors to be thrown as:
```
401 Invalid JWT
```
We now throw them as:
```
401 Unauthorized - Invalid JWT
```
(ie. status code, status error, and returned error message)

We expect to receive a message in the event of an error but if no message was returned we throw:
```
401 Unauthorized
```