Do not create SROC bill run if one already exists
https://eaflood.atlassian.net/browse/WATER-3875

The legacy system will not create a PRESROC bill run if one exists with the same region, financial year and type. There are some exceptions, for example, if the previous bill run has a status of EMPTY or ERROR.

The same logic needs to be implemented for SROC supplementary bill runs.