Add `queued` status to `LIVE_STATUSES`
We started this work on an incorrect, outdated branch. When we realised our mistake, we created a new branch and cherry-picked the commits across to it. However we'd overlooked the fact that we recently added the additional status `queued`, which we should also consider to be "live". We therefore update our `LIVE_STATUSES` list to include it