Do not create SROC bill run if one already exists (#98)
https://eaflood.atlassian.net/browse/WATER-3875

The legacy system will not create a presroc bill run if one exists that is "live" (ie. has a status of `processing`, `ready` or `review`) with the same region, financial year and type. The same logic needs to be implemented for sroc supplementary bill runs (adding our new `queued` status to the list of statuses we consider to be "live").

We do this by creating `CheckLiveBillRunService`, which takes a region id and financial year and returns a boolean to indicate whether there are any "live" bill runs. We then call this from `InitiateBillingBatchService` and throw an error if it returns `true`. Note that we give the error the same message that the `water-abstraction-service` has but we don't format it as a `409` error.