SRoC Supplementary Billing - Determine Abstraction Period (#97)
https://eaflood.atlassian.net/browse/WATER-3834

As part of determining the transactions and Billable Days for Supplementary Billing, the Abstraction Period needs to be taken into account.  Abstraction period is the date range in which Licence Holders can abstract water, in terms of Supplementary Billing, then it needs to be taken into account days when the Licence Holder cannot abstract as the dates are outside of the Abstraction Period and therefore not chargeable.

This PR is to determine the correct number of billable days in the abstraction period.