Refactor steps into functions
We've got to add another 'thing' into the service which is updating a bill run as errored. That means they'll be even more going on in the `go()` function so it felt right to start shifting some of the 'steps' out to there own functions.

The first is the step where we check if an existing bill run is already in progress.