Refactor the GlobalNotifierLib constructor
Got started on the plugin and realised we have both a Pino and Airbrake instance available on the server object.

So, we can expect both to be passed to the constructor. But to ensure we are getting existing instances rather than generating our own through BaseNotifierLib we add some logic to throw an error if any of the args are `null`.