Create `error` bill run when CM fails (#104)
https://eaflood.atlassian.net/browse/WATER-3833

When we currently initiate a new bill run, a fail response from the Charging Module means we throw an error and no bill run is created. We want to revise this so that it _does_ create a new bill run but with the status `error` and with `errorCode` set to `50`.

As part of this we update `BillingBatchModel` with a static getter `errorCodes`, which allows us to consistently use the correct error codes by referring to (for example) `BillingBatchModel.errorCodes.failedToCreateBillRun`.