Determine other factors for SROC transaction line
https://eaflood.atlassian.net/browse/WATER-3895

To get the Charging Module API to generate a charge we need more information than just billable days (see WATER-3894)

Using how the legacy code generates an SROC transaction line for an annual bill run, we need to create the same type of record in billing_transactions

So, the individual data points need to be sourced from the same places. We then need to populate the same fields.

N.B. We looking at the current SROC transactions specifically because there are fields they don't bother to populate.