Create service to consolidate date ranges (#110)
https://eaflood.atlassian.net/browse/WATER-3895

As part of our work to create transaction lines for sroc, we need to be able to take a set of dates ranges and consolidate them into a set of non-overlapping ranges. This PR implements `ConsolidateDateRangesService` to do this.

The service takes an array of date ranges, where each date range is an object with a `startDate` and `endDate` (each of which is a regular `Date` object). For example:
```js
[
  { startDate: 2023-01-01, endDate: 2023-03-01 }, // Range 1
  { startDate: 2023-02-01, endDate: 2023-04-01 }, // Range 2
  { startDate: 2023-06-01, endDate: 2023-07-01 }  // Range 3
]
```

It merges any overlapping date ranges, leaves alone any which don't overlap, and returns an array of the resulting ranges. So the date ranges above consolidate to:
```js
[
  { startDate: 2023-01-01, endDate: 2023-04-01 }, // Ranges 1 & 2 merged
  { startDate: 2023-06-01, endDate: 2023-07-01 }  // Range 3 unchanged
]
```

Note that if a range starts on the same day the previous range ends, that is classed as one continuous range. But a range that starts the day _after_ the previous range ends is classed as a separate range to that previous one. For example:

```js
[
  { startDate: 2023-01-01, endDate: 2023-04-01 }, // Range 1, ending on 1st April
  { startDate: 2023-04-01, endDate: 2023-07-01 }, // Range 2, starting on 1st April
  { startDate: 2023-10-01, endDate: 2023-10-31 }, // Range 3, ending on 31st October
  { startDate: 2023-11-01, endDate: 2023-12-01 }  // Range 4, starting on 1st November
]
```

Consolidates to:
```js
[
  { startDate: 2023-01-01, endDate: 2023-07-01 }, // Ranges 1 & 2 merged
  { startDate: 2023-10-01, endDate: 2023-10-31 }, // Range 3 unchanged
  { startDate: 2023-11-01, endDate: 2023-12-01 }  // Range 4 unchanged
]
```

One last thing to note is that, due to the way the merging has been implemented, the returned array of date ranges is sorted in order from earliest start date to latest start date.