Add SROC Supplementary Billing Invoice Service
https://eaflood.atlassian.net/browse/WATER-3896

In order to 'piggy-back' onto the existing functions in the legacy service for reviewing, confirming, sending, viewing, deleting a bill run etc we need to create the same base database records as it does for a Bill run.

This change will add a service to handle creating the `water.billing_batch_invoice` records in the DB. This is the link between the `billing_batch` and the invoice account, and from there to the licences included in the bill run.

We'll need to dig into the legacy code to understand where the previous team extracts the data that forms the `billing_invoice` records. We'll then aim to replicate that here.