Add transaction service to water-abstraction-system (#109)
https://eaflood.atlassian.net/browse/WATER-3894

Thanks to https://github.com/DEFRA/water-abstraction-system/pull/97 we now know what the billable days are for each abstraction period (`charge_purpose`) linked to a charge reference (`charge_element` && `billing_charge_categories`) within a charge version.

The final step is to take that information and start generating the transaction line information. For each charge reference in a charge version, there should be one transaction line. So, if a charge reference is linked to multiple abstraction periods, we need the sum of their billable days.