Generate Billing Invoice Licence record for SROC
https://eaflood.atlassian.net/browse/WATER-3897

In order to 'piggy-back' onto the existing functions in the legacy service for reviewing, confirming, sending, viewing, deleting a bill run etc we need to create the same base database records as it does for a Bill run.'

We already have billing_batches and events and are working on the billing_transactions in https://eaflood.atlassian.net/browse/WATER-3895 and https://eaflood.atlassian.net/browse/WATER-3894

We are also planning on building the billing_invoice record in https://eaflood.atlassian.net/browse/WATER-3896

So, the next entity we need is the billing_invoice_licence.

This is the link between the billing_batch, the invoice account, and from there to the transactions to be included in the bill run.

We'll need to dig into the legacy code to understand where the previous team extracts the data that forms the billing_invoice_licence records.