Generate Billing Invoice Licence record for SROC (#124)
https://eaflood.atlassian.net/browse/WATER-3897

In order to 'piggy-back' onto the existing functions in the legacy service for reviewing, confirming, sending, viewing, deleting a bill run etc we need to create the same base database records as it does for a Bill run. As part of this we need to generate`billing_invoice_licence` entities, which link billing invoices with licences. We already have this entity so this PR implements `CreateBillingInvoiceLicenceService` which can create these.