Create `ChargingModuleRequestLib`
https://eaflood.atlassian.net/browse/WATER-3903

Currently we have RequestLib which we use to make requests to external services. Having developed code which sends requests to the Charging Module we can see that there will be things we need to do which will be common to all CM requests -- for example, obtaining an access token, formatting the response in a particular way, etc.

Before we go any further with developing more services which send requests to the CM we want to create a general ChargingModuleRequestLib which will do all of this for us, allowing our services to concentrate on what they actually need to do and not all the stuff surrounding the requests.