Create `GlobalHapiServerMethodsPlugin`
We want to be able to access our `getChargingModuleToken()` server method in order to get and cache a Cognito token. However this means we need to pass something around from one service to another until it reaches our lib -- not ideal given that the CM request could be several services deep (plus if we pass a token around it could potentially expire). We therefore implement a `GlobalHapiServerMethodsPlugin` which lets us access them globally as `HapiServerMethods`, eg. `global.HapiServerMethods.getChargingModuleToken()`