Create `ChargingModuleRequestLib` (#129)
https://eaflood.atlassian.net/browse/WATER-3903

Currently we have `RequestLib` which we use to make requests to external services. Having developed code which sends requests to the Charging Module we can see that there will be things we need to do which will be common to all CM requests -- for example, obtaining an access token, formatting the response in a particular way, etc.

Before we go any further with developing more services which send requests to the CM we want to create a general `ChargingModuleRequestLib` which will do all of this for us, allowing our services to concentrate on what they actually need to do and not all the stuff surrounding the requests.

This PR implements `ChargingModuleRequestLib` and two initial methods, `get` and `post`.

As part of this, we want to access our server method `getChargingModuleToken()` to ensure that Cognito tokens are cached for all requests. In order to do this without passing the `server` object around from one service to another, we introduce a new `GlobalHapiServerMethodsPlugin`, which adds the Hapi server methods object to the global object, allowing us to access them from anywhere; this means we can get a token by calling `global.HapiServerMethods.getChargingModuleToken()`. In testing we confirmed that calling this and calling `server.methods.getChargingModuleToken()` gives the same token; however, for consistency we intend to use `global.HapiServerMethods.getChargingModuleToken()` at all times.

Note that existing code which sends requests to the CM will be updated in a separate PR.